Rename `sendTargetingParams.ts` to `sendTargetingParams.apps.ts`
Also, add a comment in the `getTargetingParams` logic for clarity.
