+++ dotcom-rendering/src/components/SendTargetingParams.importable.tsx
@@ -2,7 +2,7 @@ import { log } from '@guardian/libs';
 import { useEffect } from 'react';
 import { getAnalyticsClient } from '../lib/bridgetApi';
 import { isServer } from '../lib/isServer';
-import { getTargetingParams } from '../lib/sendTargetingParams';
+import { getTargetingParams } from '../lib/sendTargetingParams.apps';
 import type { EditionCommercialProperties } from '../types/commercial';
 
 type Props = {
