+++ dotcom-rendering/src/lib/sendTargetingParams.apps.ts
@@ -6,6 +6,8 @@ export const getTargetingParams = (
 ): Map<string, string> => {
 	return new Map(
 		editionCommercialProperties.adTargeting
+			// We're filtering this out because the targeting params AR sends
+			// does not include this key
 			.filter((adTargetingParam) => adTargetingParam.name !== 'sh')
 			.map((adTargetingParam) => [
 				adTargetingParam.name,
